+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @express-route-cache/core
 
+## 2.1.0
+
+### Minor Changes
+
+- 05861c8: feat: support dynamic enabled callback and caching of non-GET routes with type-safe generateCacheKey helper
+
+## 2.1.0-next.0
+
+### Minor Changes
+
+- feat: support dynamic enabled callback and caching of non-GET routes with type-safe generateCacheKey helper
+
 ## 2.0.0
 
 ### Major Changes